+++ engine/resource/src/test/test_resource.cpp
@@ -939,13 +939,13 @@ TEST(RecreateTest, RecreateTest)
 volatile bool SendReloadDone = false;
 void SendReloadThread(void*)
 {
-    uint32_t msg_size = sizeof(dmResourceDDF::Reload) + sizeof(uintptr_t) + (strlen("__testrecreate__.foo") + 1);
+    uint32_t msg_size = sizeof(dmResourceDDF::Reload) + sizeof(uint64_t) + (strlen("__testrecreate__.foo") + 1);
     dmResourceDDF::Reload* reload_resources = (dmResourceDDF::Reload*) malloc(msg_size);
     reload_resources->m_Resources.m_Count = 1;
-    uintptr_t str_ofs_offset = 16; // 0x10
-    uintptr_t str_offset = 24;//0x18;
-    memcpy((uint8_t*)reload_resources, &str_ofs_offset, sizeof(uintptr_t)); // offset to path string offsets
-    memcpy((uint8_t*)reload_resources + str_ofs_offset, &str_offset, sizeof(uintptr_t)); // offset to start of resource path string
+    uint64_t str_ofs_offset = 16; // 0x10
+    uint64_t str_offset = 24;//0x18;
+    memcpy((uint8_t*)reload_resources, &str_ofs_offset, sizeof(uint64_t)); // offset to path string offsets
+    memcpy((uint8_t*)reload_resources + str_ofs_offset, &str_offset, sizeof(uint64_t)); // offset to start of resource path string
     memcpy((uint8_t*)(reload_resources) + str_offset, "__testrecreate__.foo", strlen("__testrecreate__.foo") + 1); // the actual resource path
 
     dmMessage::URL url;
